+++ tests/Unit/Public/Get-IdoItConstant.tests.ps1
@@ -1,0 +1,32 @@
+BeforeAll {
+    $script:ModuleName = 'PSIdoitNG'
+    Remove-Module -Name $script:moduleName -Force -ErrorAction SilentlyContinue
+    Import-Module -Name $script:moduleName -Force
+
+    $PSDefaultParameterValues['InModuleScope:ModuleName'] = $script:moduleName
+    $PSDefaultParameterValues['Mock:ModuleName'] = $script:moduleName
+    $PSDefaultParameterValues['Should:ModuleName'] = $script:moduleName
+
+    $testRoot = Join-Path -Path (Get-SamplerAbsolutePath) -ChildPath 'tests'
+    $testHelpersPath = Join-Path -Path $testRoot -ChildPath 'Unit\Helpers'
+    . $testHelpersPath/MockConnectIdoit.ps1
+    . $testHelpersPath/MockData_idoit_constants_read.ps1
+    . $testHelpersPath/MockDefaultMockAtEnd.ps1
+}
+
+AfterAll {
+    Get-Module -Name $script:moduleName -All | Remove-Module -Force
+}
+
+Describe 'Get-IdoItConstant' {
+    It 'Should return a list of constants' {
+        $result = Get-IdoItConstant
+
+        $result | Should -Not -BeNullOrEmpty
+        $result.Type | Should -Not -Contain $null
+        $result.Name | Should -Not -Contain $null
+        $result.Value | Should -Not -Contain $null
+        # each name must be unique
+        $result.Name | Select-Object -Unique | Should -HaveCount $result.Count
+    }
+}